+++ integration-tests/middleware/pages/_middleware.ts
@@ -1,0 +1,7 @@
+import {NextResponse} from "next/server"
+
+export function middleware() {
+  const response = NextResponse.next()
+  response.headers.set("global-middleware", "true")
+  return response
+}