+++ integration-tests/middleware/pages/api/test.ts
@@ -1,0 +1,5 @@
+import {api} from "../../app/blitz-server"
+
+export default api((_req, res) => {
+  res.status(200).json({success: true})
+})