+++ packages/blitz-next/src/index-server.ts
@@ -13,7 +13,7 @@ import type {
   BlitzServerPlugin,
   Ctx as BlitzCtx,
   FirstParam,
-  Middleware,
+  RequestMiddleware,
   MiddlewareResponse,
 } from "blitz"
 import {handleRequestWithMiddleware, startWatcher, stopWatcher} from "blitz"
@@ -28,6 +28,8 @@ import {
 import {DefaultOptions, QueryClient} from "react-query"
 import {IncomingMessage, ServerResponse} from "http"
 import {withSuperJsonProps} from "./superjson"
+import {ParsedUrlQuery} from "querystring"
+import {PreviewData} from "next/types"
 
 export * from "./index-browser"
 
@@ -44,20 +46,33 @@ export type NextApiHandler = (
 ) => void | Promise<void>
 
 type SetupBlitzOptions = {
-  plugins: BlitzServerPlugin<Middleware, Ctx>[]
+  plugins: BlitzServerPlugin<RequestMiddleware, Ctx>[]
+  onError?: (err: Error) => void
 }
 
-export type BlitzGSSPHandler<TProps> = ({
+export type BlitzGSSPHandler<
+  TProps,
+  Query extends ParsedUrlQuery = ParsedUrlQuery,
+  PD extends PreviewData = PreviewData,
+> = ({
   ctx,
   req,
   res,
   ...args
-}: Parameters<GetServerSideProps<TProps>>[0] & {ctx: Ctx}) => ReturnType<GetServerSideProps<TProps>>
+}: Parameters<GetServerSideProps<TProps, Query, PD>>[0] & {ctx: Ctx}) => ReturnType<
+  GetServerSideProps<TProps, Query, PD>
+>
 
-export type BlitzGSPHandler<TProps> = ({
+export type BlitzGSPHandler<
+  TProps,
+  Query extends ParsedUrlQuery = ParsedUrlQuery,
+  PD extends PreviewData = PreviewData,
+> = ({
   ctx,
   ...args
-}: Parameters<GetStaticProps<TProps>>[0] & {ctx: Ctx}) => ReturnType<GetStaticProps<TProps>>
+}: Parameters<GetStaticProps<TProps, Query, PD>>[0] & {ctx: Ctx}) => ReturnType<
+  GetStaticProps<TProps, Query, PD>
+>
 
 export type BlitzAPIHandler = (
   req: Parameters<NextApiHandler>[0],
@@ -65,12 +80,14 @@ export type BlitzAPIHandler = (
   ctx: Ctx,
 ) => ReturnType<NextApiHandler>
 
-export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
-  const middlewares = plugins.flatMap((p) => p.middlewares)
+export const setupBlitzServer = ({plugins, onError}: SetupBlitzOptions) => {
+  const middlewares = plugins.flatMap((p) => p.requestMiddlewares)
   const contextMiddleware = plugins.flatMap((p) => p.contextMiddleware).filter(Boolean)
 
   const gSSP =
-    <TProps>(handler: BlitzGSSPHandler<TProps>): GetServerSideProps<TProps> =>
+    <TProps, Query extends ParsedUrlQuery = ParsedUrlQuery, PD extends PreviewData = PreviewData>(
+      handler: BlitzGSSPHandler<TProps, Query, PD>,
+    ): GetServerSideProps<TProps, Query, PD> =>
     async ({req, res, ...rest}) => {
       await handleRequestWithMiddleware<IncomingMessage, ServerResponse>(req, res, middlewares)
       const ctx = contextMiddleware.reduceRight(
@@ -94,12 +111,19 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       ctx.prefetchQuery = prefetchQuery
       ctx.prefetchInfiniteQuery = (...args) => prefetchQuery(...args, true)
 
-      const result = await handler({req, res, ctx, ...rest})
-      return withSuperJsonProps(withDehydratedState(result, queryClient))
+      try {
+        const result = await handler({req, res, ctx, ...rest})
+        return withSuperJsonProps(withDehydratedState(result, queryClient))
+      } catch (err: any) {
+        onError?.(err)
+        throw err
+      }
     }
 
   const gSP =
-    <TProps>(handler: BlitzGSPHandler<TProps>): GetStaticProps<TProps> =>
+    <TProps, Query extends ParsedUrlQuery = ParsedUrlQuery, PD extends PreviewData = PreviewData>(
+      handler: BlitzGSPHandler<TProps, Query, PD>,
+    ): GetStaticProps<TProps, Query, PD> =>
     async (context) => {
       const ctx = contextMiddleware.reduceRight((y, f) => (f ? f(y) : y), {} as Ctx)
       let queryClient: null | QueryClient = null
@@ -119,8 +143,13 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       ctx.prefetchQuery = prefetchQuery
       ctx.prefetchInfiniteQuery = (...args) => prefetchQuery(...args, true)
 
-      const result = await handler({...context, ctx: ctx})
-      return withSuperJsonProps(withDehydratedState(result, queryClient))
+      try {
+        const result = await handler({...context, ctx: ctx})
+        return withSuperJsonProps(withDehydratedState(result, queryClient))
+      } catch (err: any) {
+        onError?.(err)
+        throw err
+      }
     }
 
   const api =
@@ -129,7 +158,8 @@ export const setupBlitzServer = ({plugins}: SetupBlitzOptions) => {
       try {
         await handleRequestWithMiddleware(req, res, middlewares)
         return handler(req, res, res.blitzCtx)
-      } catch (error) {
+      } catch (error: any) {
+        onError?.(error)
         return res.status(400).send(error)
       }
     }