+++ packages/generator/templates/app/app/auth/mutations/forgotPassword.test.ts
@@ -1,4 +1,4 @@
-import { db } from "db"
+import db from "db"
 import { hash256 } from "@blitzjs/auth"
 import forgotPassword from "./forgotPassword"
 import previewEmail from "preview-email"