+++ packages/generator/templates/app/app/blitz-server.ts
@@ -1,13 +1,14 @@
 import { setupBlitzServer } from "@blitzjs/next"
 import { AuthServerPlugin, PrismaStorage } from "@blitzjs/auth"
-import {db} from "db"
 import { simpleRolesIsAuthorized } from "@blitzjs/auth"
+import db from "db"
+import { authConfig } from "./blitz-client"
 
 export const { gSSP, gSP, api } = setupBlitzServer({
   plugins: [
     AuthServerPlugin({
-      cookiePrefix: "__safeNameSlug__-cookie-prefix",
-      storage: PrismaStorage(db as any),
+      ...authConfig,
+      storage: PrismaStorage(db),
       isAuthorized: simpleRolesIsAuthorized,
     }),
   ],