+++ packages/generator/templates/mutations/delete__ModelName__.ts
@@ -1,5 +1,5 @@
 import {Ctx} from "blitz"
-import { db } from "db"
+import db from "db"
 import {z} from "zod"
 
 const Delete__ModelName__Input = z.object({