+++ packages/codemod/CHANGELOG.md
@@ -1,5 +1,126 @@
 # @blitzjs/codemod
 
+## 2.0.0-alpha.48
+
+### Patch Changes
+
+- 93851d90: Handle duplicate imports with Blitz upgrade-legacy codemod
+- Updated dependencies [93851d90]
+  - blitz@2.0.0-alpha.48
+  - @blitzjs/generator@2.0.0-alpha.48
+
+## 2.0.0-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [1d9804a6]
+- Updated dependencies [b405c1e8]
+- Updated dependencies [b86b569d]
+  - blitz@2.0.0-alpha.47
+  - @blitzjs/generator@2.0.0-alpha.47
+
+## 2.0.0-alpha.46
+
+### Patch Changes
+
+- aafdc5b4: Move middlewares from blitz config to blitz server with codemod
+  - @blitzjs/generator@2.0.0-alpha.46
+  - blitz@2.0.0-alpha.46
+
+## 2.0.0-alpha.45
+
+### Patch Changes
+
+- @blitzjs/generator@2.0.0-alpha.45
+- blitz@2.0.0-alpha.45
+
+## 2.0.0-alpha.44
+
+### Patch Changes
+
+- @blitzjs/generator@2.0.0-alpha.44
+- blitz@2.0.0-alpha.44
+
+## 2.0.0-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [527e48ac]
+  - blitz@2.0.0-alpha.43
+  - @blitzjs/generator@2.0.0-alpha.43
+
+## 2.0.0-alpha.42
+
+### Patch Changes
+
+- @blitzjs/generator@2.0.0-alpha.42
+- blitz@2.0.0-alpha.42
+
+## 2.0.0-alpha.41
+
+### Patch Changes
+
+- bbac7906: fix codemod for wrapping \_app arrow function & fix codemod for nested pages directory
+  - @blitzjs/generator@2.0.0-alpha.41
+  - blitz@2.0.0-alpha.41
+
+## 2.0.0-alpha.40
+
+### Patch Changes
+
+- @blitzjs/generator@2.0.0-alpha.40
+- blitz@2.0.0-alpha.40
+
+## 2.0.0-alpha.39
+
+### Patch Changes
+
+- Updated dependencies [b918055b]
+  - blitz@2.0.0-alpha.39
+  - @blitzjs/generator@2.0.0-alpha.39
+
+## 2.0.0-alpha.38
+
+### Patch Changes
+
+- 8aee25c5: getQueryClient function & queryClient codemod updates & shared plugin config
+- Updated dependencies [8aee25c5]
+  - @blitzjs/generator@2.0.0-alpha.38
+  - blitz@2.0.0-alpha.38
+
+## 2.0.0-alpha.37
+
+### Patch Changes
+
+- Updated dependencies [a80d2a8f]
+  - blitz@2.0.0-alpha.37
+  - @blitzjs/generator@2.0.0-alpha.37
+
+## 2.0.0-alpha.36
+
+### Patch Changes
+
+- 4cad9cca: Update queryClient import in codemod
+- Updated dependencies [4cad9cca]
+  - @blitzjs/generator@2.0.0-alpha.36
+  - blitz@2.0.0-alpha.36
+
+## 2.0.0-alpha.35
+
+### Patch Changes
+
+- e3750b04: codemod fixes
+- Updated dependencies [e3750b04]
+  - @blitzjs/generator@2.0.0-alpha.35
+  - blitz@2.0.0-alpha.35
+
+## 2.0.0-alpha.34
+
+### Patch Changes
+
+- @blitzjs/generator@2.0.0-alpha.34
+- blitz@2.0.0-alpha.34
+
 ## 2.0.0-alpha.33
 
 ### Patch Changes